Refactor State to not track number on read inputs
The `State` is now a function that takes the input character and only
returns the next state (`State`: `I` -> `State`). The `run_fsm` function
tracks how many characters are read from the input string.

The consequence of the change is that the state machines may no longer
move directly to an `accept` state on the first character read due to a
nuance in counting characters in `run_fsm`. Machines that read more than
one character must see the next character _before_ they can determine
whether to `accept` or keep reading.